requirements/dev.txt
-r base.txt

easydict
pyyaml
scipy
sklearn
click
jsonlines
matplotlib
pandas
